requirements-extra.txt
# WSGI server
gunicorn==20.1.0
# Plugins
ckanext-passwordless-api==1.1.0
# # Opentelemetry
# opentelemetry-sdk==1.15.0
# opentelemetry-api==1.15.0
# opentelemetry-exporter-otlp==1.15.0
# opentelemetry-distro==0.36b0
# opentelemetry-instrumentation==0.36b0
# opentelemetry-instrumentation-flask==0.36b0
# opentelemetry-instrumentation-wsgi==0.36b0
# opentelemetry-instrumentation-dbapi==0.36b0
# opentelemetry-instrumentation-logging==0.36b0
# opentelemetry-instrumentation-urllib==0.36b0
# opentelemetry-instrumentation-botocore==0.36b0
# opentelemetry-instrumentation-jinja2==0.36b0
# opentelemetry-instrumentation-psycopg2==0.36b0
# opentelemetry-instrumentation-redis==0.36b0
# opentelemetry-instrumentation-requests==0.36b0
# opentelemetry-instrumentation-sqlalchemy==0.36b0
# opentelemetry-instrumentation-urllib3==0.36b0
